command/jsonstate: Reject unsupported marks
Our JSON state format has support for sensitive marks in particular but
lacks generalized support and so cannot deal with any other marks.

Previously it just assumed that any mark must be the sensitive mark, but
now we'll reject other marks to ensure that if any new marks are added we
must consider whether and how the JSON state format should handle them.